requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.45
pymongo==4.15.5
dnspython==2.8.0
python-dotenv==1.2.1
greenlet==3.3.0
typing-extensions==4.15.0
pinecone==8.0.0
openai==2.14.0
psycopg2-binary==2.9.11
flask-cors==6.0.2